Drop the Minimongo collection
