+++ apps/meteor/client/lib/cachedCollections/MinimongoCollection.ts
@@ -1,7 +1,6 @@
 import { Mongo } from 'meteor/mongo';
-import { create } from 'zustand';
 
-import type { IDocumentMapStore } from './DocumentMapStore';
+import { createDocumentMapStore } from './DocumentMapStore';
 import { LocalCollection } from './LocalCollection';
 
 /**
@@ -15,73 +14,9 @@ export class MinimongoCollection<T extends { _id: string }> extends Mongo.Collec
 	 *
 	 * It should be used as a hook in React components to access the collection's records and methods.
 	 */
-	readonly use = create<IDocumentMapStore<T>>()((set, get) => ({
-		records: [],
-		has: (id: T['_id']) => get().records.some((record) => record._id === id),
-		get: (id: T['_id']) => get().records.find((record) => record._id === id),
-		some: (predicate: (record: T) => boolean) => get().records.some(predicate),
-		find: (predicate: (record: T) => boolean) => get().records.find(predicate),
-		findFirst: (predicate: (record: T) => boolean, comparator: (a: T, b: T) => number) =>
-			get().records.filter(predicate).sort(comparator)[0], // TODO: optimize this
-		filter: (predicate: (record: T) => boolean) => get().records.filter(predicate),
-		replaceAll: (records: T[]) => {
-			set({ records: records.map<T>(Object.freeze) });
-			this.recomputeQueries();
-		},
-		store: (doc) => {
-			set((state) => {
-				const records = [...state.records];
-				const index = records.findIndex((r) => r._id === doc._id);
-				if (index !== -1) {
-					records[index] = Object.freeze(doc);
-				} else {
-					records.push(Object.freeze(doc));
-				}
-				return { records };
-			});
-			this.recomputeQueries();
-		},
-		storeMany: (docs) => {
-			const records = [...get().records];
-
-			for (const doc of docs) {
-				const index = records.findIndex((r) => r._id === doc._id);
-				if (index !== -1) {
-					records[index] = Object.freeze(doc);
-				} else {
-					records.push(Object.freeze(doc));
-				}
-			}
-			set({ records });
-			this.recomputeQueries();
-		},
-		delete: (_id) => {
-			set((state) => {
-				const records = state.records.filter((r) => r._id !== _id);
-				return { records };
-			});
-			this.recomputeQueries();
-		},
-		update: (predicate: (record: T) => boolean, modifier: (record: T) => T) => {
-			set({
-				records: get().records.map((record) => (predicate(record) ? modifier(record) : record)),
-			});
-			this.recomputeQueries();
-		},
-		updateAsync: async (predicate: (record: T) => boolean, modifier: (record: T) => Promise<T>) => {
-			set({
-				records: await Promise.all(get().records.map((record) => (predicate(record) ? modifier(record) : record))),
-			});
-			this.recomputeQueries();
-		},
-		remove: (predicate: (record: T) => boolean) => {
-			set((state) => {
-				const records = state.records.filter((record) => !predicate(record));
-				return { records };
-			});
-			this.recomputeQueries();
-		},
-	}));
+	readonly use = createDocumentMapStore<T>({
+		onMutate: () => this.recomputeQueries(),
+	});
 
 	/**
 	 * The internal collection that manages the queries and results.
